+++ src/httpInterceptors/httpGatewayTimeout/httpGatewayTimeout.interceptor.ts
@@ -1,5 +1,5 @@
-import {Injectable, Optional, Injector, ClassProvider} from '@angular/core';
-import {HttpInterceptor, HttpHandler, HttpEvent, HTTP_INTERCEPTORS, HttpRequest} from '@angular/common/http';
+import {Injectable, Optional, Injector, ClassProvider, inject} from '@angular/core';
+import {HttpInterceptor, HttpHandler, HttpEvent, HTTP_INTERCEPTORS, HttpRequest, HttpHandlerFn} from '@angular/common/http';
 import {IGNORED_INTERCEPTORS} from '@anglr/common';
 import {Observable, Observer, catchError} from 'rxjs';
 
@@ -7,6 +7,7 @@ import {HttpGatewayTimeoutInterceptorOptions} from './httpGatewayTimeoutIntercep
 
 /**
  * HttpGatewayTimeoutInterceptor used for intercepting http responses and handling 504 statuses
+ * @deprecated Use new `httpGatewayTimeoutInterceptor` function
  */
 @Injectable()
 export class HttpGatewayTimeoutInterceptor implements HttpInterceptor
@@ -62,6 +63,7 @@ export class HttpGatewayTimeoutInterceptor implements HttpInterceptor
 
 /**
  * Provider for proper use of HttpGatewayTimeoutInterceptor, use this provider to inject this interceptor
+ * @deprecated Use new `httpGatewayTimeoutInterceptor` function
  */
 export const HTTP_GATEWAY_TIMEOUT_INTERCEPTOR_PROVIDER: ClassProvider =
 {
@@ -69,3 +71,45 @@ export const HTTP_GATEWAY_TIMEOUT_INTERCEPTOR_PROVIDER: ClassProvider =
     multi: true,
     useClass: HttpGatewayTimeoutInterceptor,
 };
+
+/**
+ * Interceptor used for intercepting http responses and handling 504 statuses
+ */
+export function httpGatewayTimeoutInterceptor(req: HttpRequest<unknown>, next: HttpHandlerFn): Observable<HttpEvent<unknown>>
+{
+    const injector = inject(Injector);
+    let options = inject(HttpGatewayTimeoutInterceptorOptions, {optional: true});
+
+    if(!(options instanceof HttpGatewayTimeoutInterceptorOptions))
+    {
+        options = new HttpGatewayTimeoutInterceptorOptions();
+    }
+
+    return next(req).pipe(catchError(err =>
+    {
+        return new Observable(observer =>
+        {
+            //client error, not response from server, or is ignored
+            if(err.error instanceof Error ||
+               req.context.get(IGNORED_INTERCEPTORS).some(itm => itm == HttpGatewayTimeoutInterceptor))
+            {
+                observer.error(err);
+                observer.complete();
+
+                return;
+            }
+
+            //if gateway timeout
+            if(err.status == 504)
+            {
+                options.action(injector, observer as Observer<unknown>);
+
+                return;
+            }
+
+            //other errors
+            observer.error(err);
+            observer.complete();
+        }) as Observable<HttpEvent<unknown>>;
+    }));
+}