+++ src/httpInterceptors/serviceUnavailable/serviceUnavailable.interceptor.ts
@@ -1,5 +1,5 @@
-import {Injectable, Optional, Injector, ClassProvider} from '@angular/core';
-import {HttpInterceptor, HttpHandler, HttpEvent, HTTP_INTERCEPTORS, HttpRequest} from '@angular/common/http';
+import {Injectable, Optional, Injector, ClassProvider, inject} from '@angular/core';
+import {HttpInterceptor, HttpHandler, HttpEvent, HTTP_INTERCEPTORS, HttpRequest, HttpHandlerFn} from '@angular/common/http';
 import {IGNORED_INTERCEPTORS} from '@anglr/common';
 import {Observable, Observer, catchError} from 'rxjs';
 
@@ -7,6 +7,7 @@ import {ServiceUnavailableInterceptorOptions} from './serviceUnavailableIntercep
 
 /**
  * ServiceUnavailableInterceptor used for intercepting http responses and handling 503 statuses
+ * @deprecated Use new `serviceUnavailableInterceptor` function
  */
 @Injectable()
 export class ServiceUnavailableInterceptor implements HttpInterceptor
@@ -62,6 +63,7 @@ export class ServiceUnavailableInterceptor implements HttpInterceptor
 
 /**
  * Provider for proper use of ServiceUnavailableInterceptor, use this provider to inject this interceptor
+ * @deprecated Use new `serviceUnavailableInterceptor` function
  */
 export const SERVICE_UNAVAILABLE_INTERCEPTOR_PROVIDER: ClassProvider =
 {
@@ -69,3 +71,45 @@ export const SERVICE_UNAVAILABLE_INTERCEPTOR_PROVIDER: ClassProvider =
     multi: true,
     useClass: ServiceUnavailableInterceptor,
 };
+
+/**
+ * Interceptor used for intercepting http responses and handling 503 statuses
+ */
+export function serviceUnavailableInterceptor(req: HttpRequest<unknown>, next: HttpHandlerFn): Observable<HttpEvent<unknown>>
+{
+    const injector = inject(Injector);
+    let options = inject(ServiceUnavailableInterceptorOptions, {optional: true});
+
+    if(!(options instanceof ServiceUnavailableInterceptorOptions))
+    {
+        options = new ServiceUnavailableInterceptorOptions();
+    }
+
+    return next(req).pipe(catchError(err =>
+    {
+        return new Observable(observer =>
+        {
+            //client error, not response from server, or is ignored
+            if(err.error instanceof Error ||
+               req.context.get(IGNORED_INTERCEPTORS).some(itm => itm == ServiceUnavailableInterceptor))
+            {
+                observer.error(err);
+                observer.complete();
+
+                return;
+            }
+
+            //if service unavailable
+            if(err.status == 503)
+            {
+                options.action(injector, observer as Observer<unknown>);
+
+                return;
+            }
+
+            //other errors
+            observer.error(err);
+            observer.complete();
+        }) as Observable<HttpEvent<unknown>>;
+    }));
+}